The hidden grasslands
