+++ Pokemon/Game.hpp
@@ -1,0 +1,15 @@
+
+
+class Player;
+
+class Game {
+
+public:
+
+Game();
+// game loop
+void gameLoop(Player &player);
+
+};
+
+
