+++ Pokemon/PokemonType.hpp
@@ -1,8 +1,13 @@
 // defining the enum for Pokemon Types
+#pragma once
+
 enum class PokemonType {
 	Grass,
     Fire,
     Water,
     Electric,
+    Bug,
+    Poison,
+    Rock,
     Normal // added for the default constructor
 };
