+++ Pokemon/Pokemonchoice.hpp
@@ -1,4 +1,5 @@
 // enum classes for the pokemon choice and type
+#pragma once
 
 enum class PokemonChoice {
     Bulbasaur = 1,
