+++ Pokemon/ProfessorOak.cpp
@@ -1,0 +1,54 @@
+#include "ProfessorOak.hpp"
+#include "Player.hpp"
+#include "Utility.hpp"
+
+#include <iostream>
+using namespace std;
+
+
+// Professor Class
+    ProfessorOak::ProfessorOak(string p_name){
+        name = p_name;
+    }
+
+    // Professor greets the player
+    void ProfessorOak::greetPlayer(Player& player) {
+        cout << name << ": First, tell me your name?" << endl;
+        getline(cin, player.name);
+        Utility::waitForEnter();
+        cout << "Hello " << player.name << "! Welcome to the world of Pokemon!" << endl;
+        cout << "I am Professor Oak, your guide on this journey." << endl;
+        cout << "Let's get started!" << endl;
+	}
+
+    void ProfessorOak::offerPokemonChoices(Player& player) {
+        cout << "You can choose one of these three pokemons to play with." << endl;
+        cout << "1. Bulbasaur (the wise plant guardian)" << endl;
+        cout << "2. Charmander (the fiery warrior)" << endl;
+        cout << "3. Squirtle (the cool-headed protector)" << endl;
+        cout << "Enter a number to choose from 1 to 3" << endl;
+
+        int choice;
+        cout << "Please enter the number for the pokemon of your choice" << endl;
+        cin >> choice;
+
+        player.ChoosePokemon(choice);
+	}
+
+    void ProfessorOak::explainMainQuest(Player &player){
+        
+        cout << "Professor Oak: Oak-ay! " << player.name << "I am about to explain you about your upcoming grand adventure." << endl;
+        cout << "Professor Oak: You see, becoming a Pokemon Master is no easy feat. It takes courage, wisdom, and a bit of luck." << endl;
+        cout << "Professor Oak: Your mission, should you choose to accept it (and trust me, you really do not have a choice) is to collect all the Pokémon Badges and conquer the Pokémon League." << endl;
+        cout << player.name << ": Wait... that sounds a lot like every other Pokémon game out there." << endl;
+        
+        cout << "Professor Oak: Shhh! Don't break the fourth wall " << player.name << "! This is serious business!" << endl;
+        cout << "Professor Oak: To achieve this, you will need to battle wild Pokémon, challenge gym leaders, and of course, keep your Pokémon healthy at the PokeCenter." << endl;
+        cout << "Professor Oak: You'll capture new Pokemon to strengthen your team. Just remember—there is a limit to how many Pokémon you can carry, so choose wisely!" << endl;
+        cout << "Professor Oak: Sounds like a walk in the park... right?" << endl;
+        cout << "Professor Oak: Hah! That is what they all say! But beware, young Trainer, the path to victory is fraught with challenges. And if you lose a battle... well, let us just say you'll be starting from square one." << endl;
+        cout << "Professor Oak: So, what do you say? Are you ready to become the next Pokémon Champion?" << endl;
+        cout << player.name << ": Ready as I will ever be, Professor!" << endl;
+        
+        cout << "Professor Oak: That is the spirit! Now, your journey begins. But first... let's just pretend I didn't forget to set up the actual game loop... Ahem, onwards!" << endl;
+    }
