+++ Pokemon/Utility.cpp
@@ -4,7 +4,7 @@
 #include <limits>
 using namespace std;
 
-void Utility :: clearConsole() {
+void Utility::clearConsole() {
 #ifdef _WIN32
     system("cls");
 #else
@@ -12,10 +12,10 @@ void Utility :: clearConsole() {
 #endif
 }
 
-void Utility :: waitForEnter() {
+void Utility::waitForEnter() {
     cin.get();
 }
 
-void Utility :: clearInputBuffer() {
+void Utility::clearInputBuffer() {
     cin.ignore(numeric_limits<streamsize>::max(), '\n');
 }
